+++ src/main/java/com/kozik/nursery/repositories/FeeRepository.java
@@ -4,6 +4,6 @@ import com.kozik.nursery.entities.Fee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface FeeRepository extends JpaRepository<Fee, Integer>{
+public interface FeeRepository extends JpaRepository<Fee, Long>{
     
 }
