+++ src/main/java/com/kozik/nursery/repositories/RecordRepository.java
@@ -4,6 +4,6 @@ import com.kozik.nursery.entities.Record;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface RecordRepository extends JpaRepository<Record, Integer>{
+public interface RecordRepository extends JpaRepository<Record, Long>{
     
 }
